Tagging v0.1.2

+++ deploy.sh
@@ -2,18 +2,26 @@
 set -e
 
 VERSION=$1
+TAG=`git tag -l | tr -d '\n'`
 
 if [ -z ${VERSION} ]; then
     echo Version not set
     exit 1
 fi
-
-git status
-echo -e "Continue? (Y/y): "
-read CONTINUE
+if git status; then
+    echo -e "Continue without committing changes? (Y/y): "
+    read CONTINUE
+else
+    CONTINUE=y
+fi
 
 if [[ "y" == ${CONTINUE} ]] || [[  "Y" == ${CONTINUE} ]]; then
-    git tag ${VERSION}
+    if [ "${TAG}" != "${VERSION}" ]; then
+        echo tagging
+        exit
+        git tag ${VERSION}
+    fi
+    exit
     git push github master --tags
 
     sudo docker build . --force-rm --rm -t chrisdlangton/h2o-flow:${VERSION} --compress && \
